+++ go2d/dimensions.go
@@ -1,11 +1,11 @@
 package go2d
 
 type Dimensions struct {
-    Width  int
-    Height int
+    Width  float64
+    Height float64
 }
 
-func NewSquareDimensions(size int) Dimensions {
+func NewSquareDimensions(size float64) Dimensions {
     return Dimensions{
         Width:  size,
         Height: size,
@@ -23,14 +23,14 @@ func (this *Dimensions) Plus(other Dimensions) Dimensions {
     }
 }
 
-func (this *Dimensions) PlusHeight(height int) Dimensions {
+func (this *Dimensions) PlusHeight(height float64) Dimensions {
     return this.Plus(Dimensions{
         Width:  0,
         Height: height,
     })
 }
 
-func (this *Dimensions) PlusWidth(width int) Dimensions {
+func (this *Dimensions) PlusWidth(width float64) Dimensions {
     return this.Plus(Dimensions{
         Width:  width,
         Height: 0,
